# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Aurigraph V11 Platform API
  description: |
    High-performance blockchain platform APIs for 2M+ TPS
    
    ## Overview
    Aurigraph V11 is a next-generation blockchain platform built with Java 21, Quarkus, and GraalVM.
    It provides high-throughput transaction processing, quantum-resistant cryptography, cross-chain bridging,
    and AI-driven optimization capabilities.
    
    ## Features
    - **Ultra-High Throughput**: 2M+ transactions per second
    - **Quantum-Resistant Security**: Post-quantum cryptography implementation
    - **Cross-Chain Bridge**: Interoperability with Ethereum, Solana, and other chains
    - **AI Optimization**: Machine learning-driven consensus and performance optimization
    - **HMS Integration**: Healthcare Management System integration
    - **Real-time Monitoring**: Comprehensive performance and health monitoring
    
    ## Performance Targets
    - **TPS**: 2,000,000+ transactions per second
    - **Latency**: Sub-millisecond transaction processing
    - **Startup**: <1 second native binary startup
    - **Memory**: <256MB native binary footprint
    
    ## Technology Stack
    - **Runtime**: Java 21 with Virtual Threads
    - **Framework**: Quarkus 3.26.2
    - **Native**: GraalVM native compilation
    - **Transport**: HTTP/2 with TLS 1.3, gRPC
    - **Consensus**: HyperRAFT++ algorithm
    - **Cryptography**: CRYSTALS-Kyber/Dilithium post-quantum algorithms
    
  version: 11.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v11/status:
    get:
      tags:
        - Platform Status
      summary: Get platform status
      description: Returns comprehensive platform health and status information
      operationId: getPlatformStatus
      responses:
        '200':
          description: Platform status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStatus'
              examples:
                healthy:
                  summary: Healthy platform status
                  value:
                    status: "HEALTHY"
                    version: "11.0.0"
                    uptimeSeconds: 3661
                    totalRequests: 1234567
                    platform: "Java 21 + Quarkus + GraalVM Native"
                    targetTPS: 2000000
                    timestamp: 1706275200000
        '503':
          description: Platform is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v11/crypto/sign:
    post:
      tags:
        - Security
      summary: Sign data with quantum-resistant cryptography
      description: Sign data using post-quantum digital signatures
      operationId: signData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigningRequest'
      responses:
        '200':
          description: Data signed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigningResponse'

  # ==================== BRIDGE APIs ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    PlatformStatus:
      type: object
      properties:
        status:
          type: string
          enum: [HEALTHY, WARNING, CRITICAL, MAINTENANCE]
          description: Overall platform health status
        version:
          type: string
          description: Platform version
        uptimeSeconds:
          type: integer
          format: int64
          description: Platform uptime in seconds
        totalRequests:
          type: integer
          format: int64
          description: Total number of requests processed
        platform:
          type: string
          description: Platform technology stack
        targetTPS:
          type: integer
          format: int64
          description: Target transactions per second
        timestamp:
          type: integer
          format: int64
          description: Timestamp of status check
      required:
        - status
        - version
        - uptimeSeconds
        - totalRequests
        - platform
        - targetTPS
        - timestamp
...
    SigningRequest:
      type: object
      properties:
        data:
          type: string
          description: Data to be signed
        algorithm:
          type: string
          description: Signing algorithm preference
          enum: [CRYSTALS-Dilithium, SPHINCS+, DEFAULT]
      required:
        - data
    SigningResponse:
      type: object
      properties:
        signature:
          type: string
          description: Generated signature
        algorithm:
          type: string
          description: Algorithm used for signing
        publicKey:
          type: string
          description: Public key for verification
        timestamp:
          type: integer
          format: int64

    # Bridge Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        timestamp:
          type: integer
          format: int64
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - timestamp
...